Move DebuggerProxy, Enumerator, and Node into their own files
